# api/new_openapi.yaml
openapi: 3.0.0
info:
  title: PC Management API
  version: 1.0.0
  description: PC管理システムのAPI仕様
servers:
  - url: http://localhost:3001/api
    description: 開発環境
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      summary: ログイン
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/logout:
    post:
      summary: ログアウト
      tags:
        - Authentication
      responses:
        '200':
          description: ログアウト成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: リクエストが無効です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 認証が必要です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: サーバーエラーが発生しました
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWTアクセストークン
        refresh_token:
          type: string
          description: リフレッシュトークン
        expires_in:
          type: integer
          description: トークンの有効期限（秒）
...